Adding and updating functions in the io module
Moved and updated `write_txt` function from the
`vibronic/vibronic.py` file to `util/io.py` file. This
function serves the function to write the text files in
same format as Molcas/OpenMolcas.

Updated the `open_txt` function to be able to calculate
the magnitudes as well and insert them into the data
frame. Updated documentation.

Updated documentation in the `Config` class.