Initial changes for multiple norms
Made initial changes to be able to create displaced
structures for more than one normalization value.
This included changes to the `gen_delta` function
and the `Displace.gen_displaced` method.

Something to look at in the future might be a way
to create an even set of nomrlaization values with
an extra option. This might not be necessary though.