Added functionality for multiple normalization displacements
New code added to be able to generate displacments for more than one
normaliation parameter that is given. It will automatically sort
them in such a way that the 0 index remains the equilibrium
coordinates, indeces 1 through 2*nmodes are the positive and
negative displacments for the first normalization factor
forr backwards compatibility. All indeces are now,
1+ndx*nmodes*2 through 2*nmodes+ndx*nmodes*2 where ndx is the
index for the normalizaiton parameter.

All of the parameters are sorted and the ordering is reflected
in the va.conf file that is automatically generated.