Changed ZPVC.zpvc method to accept a file as the gradient and property
Added new functionality to the ZPVC.apvc method so we can
use the config file to give the gradients and property in
a CSV formatted file. This will make it easier to run the
script as the user can save the data and we simply have to
open it.

Also added a way for the user to run the calculation on
more than one atom with a new config file option. The
code will iterate over each of the atoms and calculate
the ZPVC for each of the atoms. Can look into how we can
do this without re-calculating the gradient derivatives
as these do not depend on the chosen atom.

Fixed all of the values to make sure they match the old
scripts that were developed. This included using the NIST
conversion values.