Making zpvc more readable and adding new option
Added a new required option for the configuration
file for the temperature. This way there is a file
with the dedicated information for the temperatures
that were used in the calculation.

Moved some things around in the code to make it
neater and more readable. Still have some work to
be done with this.