Adding function to create a block diagonal matrix
Added function to read multiple txt files and combine
them into one file that is block diagonal. The file
that is generated is sparse. Meaning, only the block
diagonal elements are available in the file.

With this new function the function to combine two
SF Hamiltonian files from Molcas is modified to use
the newly created functions.