Added a way to select specific normal modes when reading file
Added a way to read a Hamiltonian CSV file and only run the
calculation for a select number of normal modes. Added some
error catching to ensure that the data is correct.

Added a new test to make sure that the program will fail
when it does not find a normal mode in the CSV file when
requesting it.