Major edits to documentation in Vibronic module
Made significant edits to docstrings in the Vibronic
module along with other edits to make the code
easier to read. Added some mathematical equations
on class docstring so that one can refer to the
equations when trying to see what the code is
actually doing.

Migrated the check_size method to its own function
as this can be used in other places.