Changed functions to get triangular matrices
Added a constraint that the input matrices must
be square to get the triangular elements of the
matrix. Made the change as we believe that this
makes more mathematical sense. You cannot place
upper triangular elemets into a non-square
matrix. For our use-case it should always be a
square matrix anyways.