Implemented new multi-point derivatives into ZPVC
Used the new implementations made to compute numerical
derivatives with two, four, six, and eight points into the
ZPVC code. This makes the code more robust as one can compute
the property with an increased precision with more data
points provided to the code. This requires that the points
are in sequential order with a constant step size. Without
this the coefficients that go into the numerical
differentiation scheme have to be re-computed.

Also, added a required input to the configuration file to
ask for the number of normalization factors and the
normalization factors themselves. This is required to
correctly keep track of which displacements correspond to
the calculations with the normalization factors.